Moves Messages Into Folder / Removes Assets Folder
