+++ client/src/components/messages/ChatWindow.sass
@@ -1,4 +1,4 @@
-@import '../vars.sass'
+@import '../../vars.sass'
 
 .chat-row
     margin: 0 auto
