+++ client/src/components/messages/Conversation.sass
@@ -1,4 +1,4 @@
-@import '../vars.sass'
+@import '../../vars.sass'
 
 .inbox-top-bar, .inbox-bottom-bar
     display: flex
