+++ client/src/components/messages/Message.sass
@@ -1,4 +1,4 @@
-@import '../vars.sass'
+@import '../../vars.sass'
 
 .message-cloud
     width: auto
